test : 수익금액 반환하는 기능 테스트 추가
